requirements.txt
# LinkIn IM - deps, Python 3.8+
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
sqlalchemy>=2.0.0
flask-sqlalchemy>=3.0.0
pyjwt>=2.8.0
bcrypt>=4.0.0
python-socketio>=5.9.0
python-engineio>=4.8.0
eventlet>=0.33.0
python-dotenv>=1.0.0
pytest>=7.0.0
